Refactor and edit visual changed and rule upload functionality.
