+++ app/components/RuleHeader/RuleHeader.module.css
@@ -13,17 +13,13 @@
   padding: 8px;
 }
 
-.homeLink {
-  color: #666;
+.homeButton {
+  color: #666 !important;
   display: inline-flex;
   align-items: center;
   gap: 4px;
   padding: 4px 8px;
   background: white;
-}
-
-.homeButton {
-  composes : homeLink;
   border: 1px solid;
   border-radius: 10px;
   font-size: 16px;
@@ -57,12 +53,6 @@
   padding: 0;
   font-size: 18px;
   color: #4b4b4b;
-}
-
-.versionInfo {
-  background: #FFFFFF;
-  border-radius: 6px;
-  padding: 4px 15px;
 }
 
 @media (max-width: 767px) {
