+++ app/components/RuleViewerEditor/RuleViewerEditor.tsx
@@ -1,8 +1,8 @@
 "use client";
-import { useState, useMemo, useEffect, useRef, useCallback } from "react";
+import { useState, useMemo, useEffect, useRef } from "react";
 import type { ReactFlowInstance } from "reactflow";
 import "@gorules/jdm-editor/dist/style.css";
-import { Spin, Modal } from "antd";
+import { Spin } from "antd";
 import { ApartmentOutlined, PlayCircleOutlined, LoginOutlined, LogoutOutlined, BookOutlined } from "@ant-design/icons";
 import {
   JdmConfigProvider,
@@ -15,16 +15,11 @@ import {
   CustomNodeSpecification,
 } from "@gorules/jdm-editor";
 import { SchemaSelectProps, PanelType } from "@/app/types/jdm-editor";
-import { Scenario, Variable } from "@/app/types/scenario";
-import { downloadFileBlob } from "@/app/utils/utils";
-import { createRuleJSONWithScenarios, convertTestsToScenarios } from "@/app/utils/ruleScenariosFormat";
-import { createScenario, getScenariosByFilename } from "@/app/utils/api";
-import { logError } from "@/app/utils/logger";
 import LinkRuleComponent from "./subcomponents/LinkRuleComponent";
 import SimulatorPanel from "./subcomponents/SimulatorPanel";
 import RuleInputOutputFieldsComponent from "./subcomponents/RuleInputOutputFieldsComponent";
 import NotesComponent from "./subcomponents/NotesComponent";
-import ScenarioSelectionContent from "./subcomponents/ScenarioSelectionContent";
+import { useJSONUpload } from "@/app/hooks/useJSONUpload";
 
 interface RuleViewerEditorProps {
   jsonFilename: string;
@@ -77,134 +72,7 @@ export default function RuleViewerEditor({
     setReactFlowRef(reactFlow);
   };
 
-  const downloadJSON = (jsonData: any, filename: string) => {
-    downloadFileBlob(JSON.stringify(jsonData, null, 2), "application/json", filename);
-  };
-
-  const handleScenarioInsertion = async () => {
-    try {
-      const updatedJSON = await createRuleJSONWithScenarios(jsonFilename, ruleContent);
-      return downloadJSON(updatedJSON, jsonFilename);
-    } catch (error: any) {
-      logError("Error fetching JSON:", error);
-      throw error;
-    }
-  };
-
-  const interceptJSONDownload = async (event: any) => {
-    if (decisionGraphRef.current && event.target?.download === "graph.json") {
-      event.preventDefault();
-      try {
-        await handleScenarioInsertion();
-      } catch (error: any) {
-        logError("Error intercepting JSON download:", error);
-      }
-    }
-  };
-
-  const handleFileUpload = useCallback(
-    async (_event: any, uploadedContent: { tests?: any[] }) => {
-      if (!uploadedContent?.tests) return;
-
-      try {
-        const [existingScenarios, scenarios] = await Promise.all([
-          getScenariosByFilename(jsonFilename),
-          Promise.resolve(convertTestsToScenarios(uploadedContent.tests)),
-        ]);
-
-        const existingTitles: Set<string> = new Set(existingScenarios.map((scenario: Scenario) => scenario.title));
-        const newScenarios = scenarios.filter((scenario) => !existingTitles.has(scenario.title));
-
-        if (newScenarios.length > 0) {
-          let selectedScenarios: Scenario[] = [];
-
-          Modal.confirm({
-            title: "Import Scenarios",
-            width: 600,
-            maskClosable: false,
-            closable: false,
-            centered: true,
-            content: (
-              <ScenarioSelectionContent
-                scenarios={newScenarios}
-                onComplete={(selected) => {
-                  selectedScenarios = selected;
-                }}
-              />
-            ),
-            okText: "Import Selected",
-            cancelText: "Cancel",
-            onOk: async () => {
-              if (selectedScenarios.length > 0) {
-                const ruleId = getRuleIdFromPath();
-                await Promise.all(
-                  selectedScenarios.map((scenario) =>
-                    createScenario({ ...scenario, ruleID: ruleId, filepath: jsonFilename })
-                  )
-                ).then(() => {
-                  updateScenarios();
-                });
-              }
-            },
-          });
-        }
-      } catch (error) {
-        console.error("Failed to process scenarios:", error);
-      }
-    },
-    [jsonFilename, updateScenarios]
-  );
-
-  useEffect(() => {
-    const handleFileSelect = (event: any) => {
-      if (
-        decisionGraphRef.current &&
-        event.target?.accept === "application/json" &&
-        event.target.type === "file" &&
-        event.target.files.length > 0
-      ) {
-        const file = event.target.files[0];
-
-        // Parse contents of the uploaded file
-        const reader = new FileReader();
-        reader.onload = async (e) => {
-          try {
-            if (e.target?.result) {
-              const fileContent = JSON.parse(e.target.result as string);
-              await handleFileUpload(event, fileContent);
-            }
-          } catch (error) {
-            console.error("Error parsing JSON file:", error);
-          }
-        };
-        reader.readAsText(file);
-      }
-    };
-
-    document.addEventListener("change", handleFileSelect, true);
-
-    return () => {
-      document.removeEventListener("change", handleFileSelect, true);
-    };
-  }, [decisionGraphRef, handleFileUpload]);
-
-  useEffect(() => {
-    const clickHandler = (event: any) => {
-      interceptJSONDownload(event);
-    };
-
-    document.addEventListener("click", clickHandler);
-
-    return () => {
-      document.removeEventListener("click", clickHandler);
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [ruleContent]);
-
-  const getRuleIdFromPath = () => {
-    const match = window.location.pathname.match(/\/rule\/([^/]+)/);
-    return match?.[1] ?? null;
-  };
+  useJSONUpload(jsonFilename, updateScenarios, decisionGraphRef, ruleContent);
 
   const additionalComponents: NodeSpecification[] = useMemo(
     () => [
