+++ app/components/SavePublish/SavePublish.module.css
@@ -1,7 +1,3 @@
-.savePublishWrapper {
-  background: transparent;
-}
-
 .sendForReviewBtn {
   background: orange !important;
   border-color: orange !important;
